fastapi/requirements.txt
fastapi>=0.68.0,<0.69.0
pydantic>=1.8.0,<2.0.0
uvicorn>=0.15.0,<0.16.0
langchain==0.1.16
langchain-core==0.1.52
langchain-openai==0.1.3
python-dotenv==1.0.1
redis==5.0.3
